% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{biolizard_pal_beige_gn_blue}
\alias{biolizard_pal_beige_gn_blue}
\title{beige-green-blue continuous scale}
\usage{
biolizard_pal_beige_gn_blue(n, reverse = FALSE)
}
\arguments{
\item{n}{Integer, number of colors.}

\item{reverse}{Boolean, if TRUE, reverse the order of the colors}
}
\description{
Continuous, perceptually uniform, color-blind friendly color scale inspired by the beige, green and blue colors in the brand book.
It was created using \code{viscm} (https://github.com/matplotlib/viscm).
}
\examples{
biolizard_pal_beige_gn_blue(10)
biolizard_pal_beige_gn_blue(4, reverse = TRUE)

}
\seealso{
\code{\link[=biolizard_pal_paired]{biolizard_pal_paired()}} \code{\link[=biolizard_pal_hue]{biolizard_pal_hue()}} \code{\link[=biolizard_pal_qualitative]{biolizard_pal_qualitative()}} \code{\link[=biolizard_pal_divergent]{biolizard_pal_divergent()}} \code{\link[=biolizard_pal_beige_blue]{biolizard_pal_beige_blue()}} \code{\link[=biolizard_pal_sequential]{biolizard_pal_sequential()}}
}
