% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{biolizard_pal_divergent}
\alias{biolizard_pal_divergent}
\title{Biolizard divergent palette}
\usage{
biolizard_pal_divergent(n, reverse = FALSE)
}
\arguments{
\item{n}{Integer, number of colors.}

\item{reverse}{Boolean, if TRUE, reverse the order of the colors}
}
\description{
Divergent, colorblind-safe and perceptually uniform (within each branch) continuous color palette inspired by the green-blue color gradient in the biolizard brand book.
}
\examples{
biolizard_pal_divergent(10)
biolizard_pal_divergent(4, reverse = TRUE)

}
\seealso{
\code{\link[=biolizard_pal_paired]{biolizard_pal_paired()}} \code{\link[=biolizard_pal_hue]{biolizard_pal_hue()}} \code{\link[=biolizard_pal_qualitative]{biolizard_pal_qualitative()}} \code{\link[=biolizard_pal_sequential]{biolizard_pal_sequential()}} \code{\link[=biolizard_pal_beige_blue]{biolizard_pal_beige_blue()}} \code{\link[=biolizard_pal_beige_gn_blue]{biolizard_pal_beige_gn_blue()}}
}
