% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{biolizard_pal_hue}
\alias{biolizard_pal_hue}
\title{Biolizard hues palette}
\usage{
biolizard_pal_hue(n, reverse = FALSE)
}
\arguments{
\item{n}{Integer, number of colors.}

\item{reverse}{Boolean, if TRUE, reverse the order of the colors}
}
\description{
Specific case of the \code{scales::pal_hue} function that returns a discrete color scale
starting with the dark green color in the brand book. This scale can be expanded to large
number of colors, but DOES NOT guarantee colorblind-safe palettes.
}
\examples{
biolizard_pal_hue(12)
biolizard_pal_hue(5, reverse = TRUE)

}
\seealso{
\code{\link[=biolizard_pal_paired]{biolizard_pal_paired()}} \code{\link[=biolizard_pal_qualitative]{biolizard_pal_qualitative()}} \code{\link[=biolizard_pal_divergent]{biolizard_pal_divergent()}} \code{\link[=biolizard_pal_sequential]{biolizard_pal_sequential()}} \code{\link[=biolizard_pal_beige_blue]{biolizard_pal_beige_blue()}} \code{\link[=biolizard_pal_beige_gn_blue]{biolizard_pal_beige_gn_blue()}}
}
