% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{biolizard_pal_l_viridis}
\alias{biolizard_pal_l_viridis}
\title{l_viridis continuous color scale}
\usage{
biolizard_pal_l_viridis(n, reverse = FALSE)
}
\arguments{
\item{n}{Integer, number of colors.}

\item{reverse}{Boolean, if TRUE, reverse the order of the colors}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use biolizard_pal_beige_blue(), biolizard_pal_beige_gn_blue() or biolizard_pal_sequential() instead. Returns values of biolizard_pal_beige_gn_blue().

(deprecated) This color scale is inspired by the viridis colorscale, and passes through biolizard's signature yellow and green (or very similar).
Like viridis it is colorblind-friendly and perceptually uniform.
It was created using \code{viscm} (https://github.com/matplotlib/viscm) and named l_viridis, after the European green lizard (Lacerta viridis)
}
\examples{
biolizard_pal_l_viridis(10)
biolizard_pal_l_viridis(4, reverse = TRUE)

}
\seealso{
\code{\link[=biolizard_pal_beige_blue]{biolizard_pal_beige_blue()}} \code{\link[=biolizard_pal_beige_gn_blue]{biolizard_pal_beige_gn_blue()}} \code{\link[=biolizard_pal_sequential]{biolizard_pal_sequential()}}
}
