% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{biolizard_pal_paired}
\alias{biolizard_pal_paired}
\title{Biolizard qualitative paired palette}
\usage{
biolizard_pal_paired(n, reverse = FALSE)
}
\arguments{
\item{n}{Integer, number of colors. Maximum 10.}

\item{reverse}{Boolean, if TRUE, reverse the order of the colors}
}
\description{
Returns a colorblind-safe qualitative color palette inspired by Biolizard's signature green, blue and yellow.
Every two consecutive colors have a similar hue, making this palette especially suitable for paired levels:
e.g. two cases for different time points, such as control_t1, case_t1, control_t2, case_t2, control_t3, case_t3, ...
}
\examples{
biolizard_pal_paired(6)
biolizard_pal_paired(4, reverse = TRUE)

}
\seealso{
\code{\link[=biolizard_pal_hue]{biolizard_pal_hue()}} \code{\link[=biolizard_pal_qualitative]{biolizard_pal_qualitative()}} \code{\link[=biolizard_pal_sequential]{biolizard_pal_sequential()}} \code{\link[=biolizard_pal_divergent]{biolizard_pal_divergent()}}
}
