% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{biolizard_pal_sequential}
\alias{biolizard_pal_sequential}
\title{Biolizard sequential palette}
\usage{
biolizard_pal_sequential(n, reverse = FALSE)
}
\arguments{
\item{n}{Integer, number of colors.}

\item{reverse}{Boolean, if TRUE, reverse the order of the colors}
}
\description{
Sequential, colorblind-safe and perceptually uniform continuous color palette inspired by the light and darker shades of blue in the BioLizard brand book.
}
\examples{
biolizard_pal_sequential(10)
biolizard_pal_sequential(4, reverse = TRUE)

}
\seealso{
\code{\link[=biolizard_pal_paired]{biolizard_pal_paired()}} \code{\link[=biolizard_pal_hue]{biolizard_pal_hue()}} \code{\link[=biolizard_pal_qualitative]{biolizard_pal_qualitative()}} \code{\link[=biolizard_pal_divergent]{biolizard_pal_divergent()}} \code{\link[=biolizard_pal_beige_blue]{biolizard_pal_beige_blue()}} \code{\link[=biolizard_pal_beige_gn_blue]{biolizard_pal_beige_gn_blue()}}
}
