% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\docType{data}
\name{biolizard_qualitative_colors}
\alias{biolizard_qualitative_colors}
\title{Biolizard qualitative colors}
\format{
An object of class \code{character} of length 6.
}
\usage{
biolizard_qualitative_colors
}
\description{
A set of colors primarily chosen for discrete data representation.
}
\keyword{datasets}
