% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\docType{data}
\name{biolizard_qualitative_palette}
\alias{biolizard_qualitative_palette}
\title{Biolizard Qualitative Color Palette}
\format{
An object of class \code{character} of length 12.
}
\usage{
biolizard_qualitative_palette
}
\description{
This palette is inspired by Martin Krzywinski's "12-COLOR PALETTE FOR COLORBLINESS"
and is infused with the signature house colors of Biolizard for the first three shades.
Tailored to be inclusive, it is friendly for those with the prevalent form of color blindness: Deuteranopia (Red-Green Color Blindness).
}
\keyword{datasets}
