% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioLizardStyleR-package.R
\docType{package}
\name{BioLizardStyleR-package}
\alias{BioLizardStyleR}
\alias{BioLizardStyleR-package}
\title{BioLizardStyleR: MAKING GGPLOT GRAPHICS IN BIOLIZARD STYLE}
\description{
The BioLizardStyleR package offers tools to craft ggplot2 visualizations in the distinct BioLizard style. It features an integrated font installer, a dedicated ggplot2 theme function, a collection of custom color functions, and a function for appending a Biolizard footer and exporting the resultant graph.
}
\author{
\strong{Maintainer}: Alexander Koch \email{alexander.koch@lizard.bio}

Authors:
\itemize{
  \item Aniko Meijer \email{aniko.meijer@lizard.bio}
  \item Robbe Neirynck \email{robbe.neirynck1@hotmail.com}
}

}
\keyword{internal}
