% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_lizardplot.R
\name{finalise_lizardplot}
\alias{finalise_lizardplot}
\title{Finalize and Save Plot in BioLizard Style}
\usage{
finalise_lizardplot(
  ggplot_name,
  source = "",
  save_filepath = NULL,
  output_name = "TempLizardPlot",
  width_pixels = 640,
  height_pixels = 450,
  logo_image_path = get_image_path(),
  device = "png"
)
}
\arguments{
\item{ggplot_name}{A ggplot object.}

\item{source}{Text to display as the source in the footer.}

\item{save_filepath}{Path where the plot should be saved. Default is the current working directory.}

\item{output_name}{Optional name for the output plot file.}

\item{width_pixels}{Width of the output plot in pixels.}

\item{height_pixels}{Height of the output plot in pixels.}

\item{logo_image_path}{Path to the logo image. Defaults to the Biolizard logo within the package.}

\item{device}{The output format (e.g., "png", "pdf", "svg", "ps").}
}
\value{
Invisibly returns the styled plot.
}
\description{
Appends a Biolizard footer to a ggplot and exports the resultant graph.
It utilizes the \code{save_plot} function for saving the plot.
}
\details{
When specifying both \code{save_filepath} and \code{output_name}, note that the \code{save_filepath}
takes precedence. If \code{save_filepath} is set to its default and \code{output_name} is specified,
then the plot will be saved with the \code{output_name} in the current working directory.
}
