% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_lizardplot.R
\name{get_image_path}
\alias{get_image_path}
\title{Retrieve Path to Image in the Package}
\usage{
get_image_path()
}
\value{
Full path to the image file.
}
\description{
Returns the system file path to the Biolizard logo
}
