% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_.biolizard_fonts.R
\name{get_path}
\alias{get_path}
\title{Retrieve the Absolute Path to the Fonts Folder}
\usage{
get_path()
}
\value{
A character string representing the absolute path to the fonts folder.
}
\description{
This function returns the absolute path to the 'fonts' folder within the 'BioLizardStyleR' package.
}
\examples{
get_path()
}
