% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_biolizard_fonts.R
\name{install_biolizard_fonts}
\alias{install_biolizard_fonts}
\title{Import BioLizard Fonts into R}
\usage{
install_biolizard_fonts(
  pattern = "[Rr]ed [[Hh]at [Dd]isplay",
  clearDataBase = FALSE
)
}
\arguments{
\item{pattern}{A character string specifying the font file pattern. Default is 'Red Hat Display'.}

\item{clearDataBase}{A logical flag to clear the existing database of fonts in 'extrafont'. Useful if pre-existing fonts are causing issues. Default is FALSE.}
}
\value{
Prints a success message if the desired font is loaded correctly. Additionally, warnings are issued if multiple fonts match the pattern or if no matching fonts are found.
}
\description{
This function imports BioLizard fonts into R using 'Red Hat Display' as the default font.
Users with a different Red Hat Display font installed can adjust the pattern. The function leverages the 'extrafont'
package and provides feedback on fonts that match the specified pattern.
}
\examples{
\dontrun{
install_biolizard_fonts()
extrafont::fonts()  # to check the installed fonts
}
}
