% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_lizardplot.R
\name{left_align}
\alias{left_align}
\title{Left Align Text in ggplot}
\usage{
left_align(ggplot_name, pieces)
}
\arguments{
\item{ggplot_name}{A ggplot object.}

\item{pieces}{A character vector of ggplot text elements (e.g., "title", "subtitle", "caption").}
}
\value{
A grob object with left-aligned text.
}
\description{
Aligns specified text elements of a ggplot to the left.
}
