% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lizard_style.R
\name{lizard_layout}
\alias{lizard_layout}
\title{Apply the lizard layout to a plotly plot}
\usage{
lizard_layout(fig, ...)
}
\arguments{
\item{fig}{A plotly figure created with standard plotly language.}

\item{...}{Further arguments for the plotly::layout() function}
}
\value{
A plotly figure in the BioLizard style
}
\description{
This function applies a predefined 'Lizard' style to a plotly plot. It sets specific font types,
sizes and other graphical elements to ensure the plot conforms to the BioLizard brand.
}
\details{
Adapted from: https://github.com/plotly/plotly.R/issues/2117
}
\examples{
library(plotly)
# Works with plotly
plot_ly(
  mtcars,
  type = "scatter",
  mode = "markers",
  x = ~mpg,
  y = ~disp,
  color = ~gear,
  colors = biolizard_pal_l_viridis(4)) |> lizard_layout()

# Works with ggplotly
p <- ggplot(mtcars, aes(mpg, disp)) + geom_point()
ggplotly(p) |> lizard_layout()
}
