% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_colors.R
\name{reset_colors}
\alias{reset_colors}
\title{Reset default ggplot2 colors}
\usage{
reset_colors()
}
\description{
This function resets the default colors for elements of ggplot2 plots, that
are changed to BLZ green ("#01a086") when using the \code{lizard_style()} function.
Note that colors are automatically reset when restarting the R session.
}
\examples{
library(ggplot2)
\dontrun{p <- ggplot(mtcars, aes(mpg, disp)) + geom_point()
p + lizard_style()
p  # default colors have changed to BLZ green, even without applying the theme
reset_colors()
p  # ggplot default colors
}
}
