% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_lizardplot.R
\name{save_plot}
\alias{save_plot}
\title{Save a ggplot2 Plot to File}
\usage{
save_plot(plot_grid, width, height, save_filepath, device)
}
\arguments{
\item{plot_grid}{The ggplot2 plot to be saved.}

\item{width}{Width of the output image in pixels.}

\item{height}{Height of the output image in pixels.}

\item{save_filepath}{Path where the plot should be saved.}

\item{device}{The output format (e.g., "png", "pdf", "svg", "ps").}
}
\description{
This function saves a given ggplot2 plot to a specified filepath using
the ggplot2 package for PNGs and the Cairo package for other formats.
}
