% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_color_gradient_biolizard}
\alias{scale_color_gradient_biolizard}
\alias{scale_colour_gradient_biolizard}
\title{Biolizard gradient color scale}
\usage{
scale_color_gradient_biolizard(...)

scale_colour_gradient_biolizard(...)
}
\description{
Biolizard gradient color scale

Alternative spelling for gradient color scale
}
