% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_color_sequential_continuous_biolizard}
\alias{scale_color_sequential_continuous_biolizard}
\alias{scale_colour_sequential_continuous_biolizard}
\title{Sequential Continuous Biolizard Color Scale}
\usage{
scale_color_sequential_continuous_biolizard(...)

scale_colour_sequential_continuous_biolizard(...)
}
\description{
Use this function with ggplot when fill corresponds to a continuous range, applying the sequential Biolizard palette.
}
