% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_colour_qualitative_biolizard}
\alias{scale_colour_qualitative_biolizard}
\alias{scale_color_qualitative_biolizard}
\title{Biolizard qualitative color scale}
\usage{
scale_colour_qualitative_biolizard(...)

scale_color_qualitative_biolizard(...)
}
\description{
Biolizard qualitative color scale

Alternative spelling for color scale
}
