% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_colour_sequential_biolizard}
\alias{scale_colour_sequential_biolizard}
\alias{scale_color_sequential_biolizard}
\title{Biolizard sequential color scale}
\usage{
scale_colour_sequential_biolizard(...)

scale_color_sequential_biolizard(...)
}
\description{
Provides a function for the sequential color scale.
}
