% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_fill_divergent_discrete_biolizard}
\alias{scale_fill_divergent_discrete_biolizard}
\title{Divergent Discrete Biolizard Fill Scale}
\usage{
scale_fill_divergent_discrete_biolizard(...)
}
\description{
Use this function with ggplot when fill corresponds to discrete categories with a natural midpoint, applying the divergent Biolizard palette.
}
\details{
This divergent palette codes underlying numeric values by a triangular luminance sequence with different hues in the left and in the right “arms” of the palette.
(a) a single hue is used for each arm of the palette,
(b) chroma and luminance trajectory are balanced between the two arms,
(c) the neutral central value has zero chroma.
The palette is crafted using hue 291 and hue 170, which is the distinctive biolizard green. .
This unique hue pairing produces a palette that remains accessible for all major forms of color blindness.
}
