% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_fill_gradient_biolizard}
\alias{scale_fill_gradient_biolizard}
\title{Biolizard gradient fill scale}
\usage{
scale_fill_gradient_biolizard(...)
}
\description{
Biolizard gradient fill scale
}
