% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_fill_qualitative_biolizard}
\alias{scale_fill_qualitative_biolizard}
\title{Biolizard qualitative fill scale}
\usage{
scale_fill_qualitative_biolizard(...)
}
\description{
Biolizard qualitative fill scale
}
