% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_fill_sequential_biolizard}
\alias{scale_fill_sequential_biolizard}
\title{Biolizard sequential fill scale}
\usage{
scale_fill_sequential_biolizard(...)
}
\description{
Biolizard sequential fill scale
}
