% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lizard_style.R
\name{set_default_BLZcolors}
\alias{set_default_BLZcolors}
\title{set default colors of plot elements to black}
\usage{
set_default_BLZcolors()
}
\description{
set default colors of plot elements to black
}
