% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lizard_style.R
\name{set_default_BLZgreen}
\alias{set_default_BLZgreen}
\title{set default colors of plot elements to BLZ green}
\usage{
set_default_BLZgreen()
}
\description{
set default colors of plot elements to BLZ green
}
