% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{biolizard_pal_divergent}
\alias{biolizard_pal_divergent}
\title{Biolizard divergent palette}
\usage{
biolizard_pal_divergent(n, reverse = FALSE)
}
\arguments{
\item{n}{Integer, number of colors.}

\item{reverse}{Boolean, if TRUE, reverse the order of the colors}
}
\description{
Divergent color palette inspired by Biolizard's signature green.
}
\examples{
biolizard_pal_divergent(10)
biolizard_pal_divergent(4, reverse = TRUE)

}
\seealso{
\code{\link{biolizard_pal_paired}} \code{\link{biolizard_pal_hue}} \code{\link{biolizard_pal_qualitative}} \code{\link{biolizard_pal_sequential}}
}
