% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{biolizard_pal_l_viridis}
\alias{biolizard_pal_l_viridis}
\title{l_viridis continuous color scale}
\usage{
biolizard_pal_l_viridis(n, reverse = FALSE)
}
\arguments{
\item{n}{Integer, number of colors.}

\item{reverse}{Boolean, if TRUE, reverse the order of the colors}
}
\description{
This color scale is inspired by the viridis colorscale, and passes through biolizard's signature yellow and green (or very similar).
Like viridis it is colorblind-friendly and perceptually uniform.
It was created using \code{viscm} (https://github.com/matplotlib/viscm) and named l_viridis, after the European green lizard (Lacerta viridis)
}
\examples{
biolizard_pal_l_viridis(10)
biolizard_pal_l_viridis(4, reverse = TRUE)

}
\seealso{
\code{\link[=biolizard_pal_paired]{biolizard_pal_paired()}} \code{\link[=biolizard_pal_hue]{biolizard_pal_hue()}} \code{\link[=biolizard_pal_qualitative]{biolizard_pal_qualitative()}} \code{\link[=biolizard_pal_divergent]{biolizard_pal_divergent()}}
}
