% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{biolizard_pal_qualitative}
\alias{biolizard_pal_qualitative}
\title{Biolizard qualitative 12-color palette}
\usage{
biolizard_pal_qualitative(n, reverse = FALSE)
}
\arguments{
\item{n}{Integer, number of colors. Maximum 12.}

\item{reverse}{Boolean, if TRUE, reverse the order of the colors}
}
\description{
Returns a colorblind-safe qualitative color palette starting with Biolizard's signature green, blue and yellow.
}
\examples{
biolizard_pal_qualitative(8)
biolizard_pal_qualitative(4, reverse = TRUE)

}
\seealso{
\code{\link{biolizard_pal_paired}} \code{\link{biolizard_pal_hue}} \code{\link{biolizard_pal_sequential}} \code{\link{biolizard_pal_divergent}}
}
