% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{biolizard_pal_qualitative}
\alias{biolizard_pal_qualitative}
\title{Biolizard qualitative color palette function}
\usage{
biolizard_pal_qualitative()
}
\value{
A function that takes an integer \code{n} and returns a vector of \code{n} colors.
}
\description{
Provides a function to retrieve a specified number of qualitative colors from the Biolizard palette.
}
