% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{biolizard_pal_sequential}
\alias{biolizard_pal_sequential}
\title{Biolizard sequential color palette function}
\usage{
biolizard_pal_sequential()
}
\value{
A function that takes an integer \code{n} and returns a vector of \code{n} colors.
}
\description{
Provides a function that returns a specified number of sequential colors using the cividis scale.
}
