% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_color_biolizard}
\alias{scale_color_biolizard}
\alias{scale_colour_biolizard}
\alias{scale_fill_biolizard}
\title{Biolizard ggplot2 Scales}
\usage{
scale_color_biolizard(
  type = "discrete",
  scheme = "qualitative",
  reverse = FALSE,
  ...
)

scale_colour_biolizard(
  type = "discrete",
  scheme = "qualitative",
  reverse = FALSE,
  ...
)

scale_fill_biolizard(
  type = "discrete",
  scheme = "qualitative",
  reverse = FALSE,
  ...
)
}
\arguments{
\item{type}{Character, either 'discrete' or 'continuous'.}

\item{scheme}{Character, either 'qualitative', 'sequential', or 'divergent'.}

\item{reverse}{Boolean, if TRUE, reverse the order of the colors in the palette.}

\item{...}{Additional arguments passed on to the underlying ggplot2 scale functions. For more details on these arguments, refer to the standard ggplot2 scales documentation. For discrete type, refer to the documentation of \code{\link[ggplot2]{discrete_scale}}. For continuous type, refer to the documentation of \code{\link[ggplot2]{scale_fill_gradientn}}. Two useful examples for continuous maps are trans='reverse' or limits=c(-1,1).}
}
\description{
Use this function to incorporate the Biolizard palettes as a color scale into a ggplot object.

The function supports three schemes:

\itemize{
  \item \strong{qualitative}: Colorblind-safe qualitative color palette starting with Biolizard's signature green, blue and yellow.
  \item \strong{paired}: Colorblind-safe qualitative color palette inspired by Biolizard's signature colors, especially suited for levels that are related 2-by-2.
  \item \strong{hues}: Maps each level to an evenly spaced hue on the color wheel, starting with Biolizard's signature green. DOES NOT generate colorblind-safe palettes.
  \item \strong{sequential}: Sequential color palette inspired by Biolizard's signature green.
  \item \strong{divergent}: Divergent color palette inspired by Biolizard's signature green.
}
}
\section{Application}{

Add the \code{scale_color_biolizard()} or \code{scale_fill_biolizard()} function to your ggplot chain to apply the respective color or fill palette.
}

\examples{
\dontrun{
library(ggplot2)
data("mtcars")
mtcars$gear <- as.factor(mtcars$gear)
testplot <- ggplot(data = mtcars, aes(x = hp, y = mpg)) +
 geom_point(aes(color = gear),size=3) +
 labs(title = "Miles per Gallon vs. Horsepower",
      x = "Horsepower",
      y = "Miles per Gallon",
      color = "Gears")
 testplot +  scale_color_biolizard(type='discrete',scheme='qualitative')
}

}
\seealso{
\code{\link{biolizard_pal_paired}} \code{\link{biolizard_pal_hue}} \code{\link{biolizard_pal_qualitative}} \code{\link{biolizard_pal_sequential}} \code{\link{biolizard_pal_divergent}}
}
