% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\docType{data}
\name{blz_blue}
\alias{blz_blue}
\title{Biolizard dark blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
blz_blue
}
\description{
Biolizard dark blue
}
\keyword{datasets}
