% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{blz_green}
\alias{blz_green}
\title{Biolizard green}
\usage{
blz_green
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The BioLizard green (#01a086) is no longer part of the BioLizard brand.
This variable now returns black (#000000) to prevent breaking code.
}
