% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\docType{data}
\name{blz_green}
\alias{blz_green}
\title{Biolizard green}
\format{
An object of class \code{character} of length 1.
}
\usage{
blz_green
}
\description{
Biolizard green
}
\keyword{datasets}
