% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\docType{data}
\name{blz_yellow}
\alias{blz_yellow}
\title{Biolizard yellow}
\format{
An object of class \code{character} of length 1.
}
\usage{
blz_yellow
}
\description{
Biolizard yellow
}
\keyword{datasets}
