% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_lizardplot.R
\name{create_footer}
\alias{create_footer}
\title{Create Footer with Dynamic Font Size, Logo, and Custom Font}
\usage{
create_footer(source, logo_image_path = get_image_path())
}
\arguments{
\item{source}{Text to display as the source in the footer.}

\item{logo_image_path}{Path to the logo image. Defaults to the "BiolizardLogo.png" within the package.}
}
\value{
A grob object representing the footer with the source text, logo, and the specified font.
}
\description{
Constructs a footer for a ggplot using specified source text, a logo image,
and an optional font. The font size of the source text is determined dynamically
based on its length using the \code{\link{compute_font_size}} function.
}
