% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_lizardplot.R
\name{finalise_lizardplot}
\alias{finalise_lizardplot}
\title{Finalize and Save Plot in BioLizard Style}
\usage{
finalise_lizardplot(
  ggplot_name,
  source,
  save_filepath = paste0(gsub("\\\\\\\\", "/", getwd()), "/TempLizardPlot.png"),
  output_name = NULL,
  width_pixels = 640,
  height_pixels = 450,
  logo_image_path = get_image_path()
)
}
\arguments{
\item{ggplot_name}{A ggplot object.}

\item{source}{Text to display as the source in the footer.}

\item{save_filepath}{Path where the plot should be saved. Default is current working directory.}

\item{output_name}{Optional name for the output plot file.}

\item{width_pixels}{Width of the output plot in pixels.}

\item{height_pixels}{Height of the output plot in pixels.}

\item{logo_image_path}{Path to the logo image. Defaults to the Biolizard logo within the package.}
}
\value{
Invisibly returns the styled plot.
}
\description{
Appends a Biolizard footer and exports the resultant graph
}
\details{
When specifying both `save_filepath` and `output_name`, note that the `save_filepath`
takes precedence. If `save_filepath` is set to its default and `output_name` is specified,
then the plot will be saved with the `output_name` in the current working directory.
}
