% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_.biolizard_fonts.R
\name{install_biolizard_fonts}
\alias{install_biolizard_fonts}
\title{Import BioLizard Fonts into R}
\usage{
install_biolizard_fonts(pattern = "AvenirLTStd-Roman.ttf")
}
\arguments{
\item{pattern}{A character string specifying the font file pattern. Default is 'AvenirLTStd-Roman.ttf'.}
}
\value{
Prints a success message if the desired font is loaded correctly. Additionally, warnings are issued if multiple fonts match the pattern or if no matching fonts are found.
}
\description{
This function imports BioLizard fonts into R using 'AvenirLTStd-Roman.ttf' as the default font.
Users with a different Avenir font installed can adjust the pattern. A simple starting option
might be 'Avenir' if the exact .ttf filename is unknown. The function leverages the 'extrafont'
package and provides feedback on fonts that match the specified pattern.
}
\examples{
\dontrun{
install_biolizard_fonts()
extrafont::fonts()  # to check the installed fonts
}
}
