% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_.biolizard_fonts.R
\name{install_biolizard_fonts}
\alias{install_biolizard_fonts}
\title{Import BioLizard Fonts into R}
\usage{
install_biolizard_fonts()
}
\description{
This function imports the fonts from the 'BioLizardStyleR' package into R using the 'extrafont' package.
}
\examples{
install_biolizard_fonts()
extrafont::fonts()  # to check the installed fonts
}
