% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lizard_style.R
\name{lizard_style}
\alias{lizard_style}
\title{Apply the lizard Style Theme to a ggplot2 Plot}
\usage{
lizard_style()
}
\value{
A \code{ggplot2::theme} object that can be added to a ggplot2 plot.
}
\description{
This function applies a predefined 'Lizard' style to a ggplot2 plot. It sets specific font types,
sizes and other graphical elements to ensure the plot conforms to the BioLizard brand book.
}
\examples{
\dontrun{library(ggplot2)
p <- ggplot(mtcars, aes(mpg, disp)) + geom_point()
p + lizard_style()
}
}
