% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_colour_qualitative_biolizard}
\alias{scale_colour_qualitative_biolizard}
\alias{scale_color_qualitative_biolizard}
\title{Qualitative Biolizard Color Scale}
\usage{
scale_colour_qualitative_biolizard(...)

scale_color_qualitative_biolizard(...)
}
\description{
Use this function to incorporate the [biolizard_qualitative_palette] as a color scale into a ggplot object.
}
\seealso{
[biolizard_qualitative_palette] for the detailed palette description and colors.
}
