% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_functions.R
\name{scale_fill_qualitative_biolizard}
\alias{scale_fill_qualitative_biolizard}
\title{Qualitative Biolizard Fill Scale}
\usage{
scale_fill_qualitative_biolizard(...)
}
\description{
Use this function to incorporate the [biolizard_qualitative_palette] as a fill scale into a ggplot object.
}
\seealso{
[biolizard_qualitative_palette] for the detailed palette description and colors.
}
